Split up source into modules
