+++ src/oled_monitor/interrupt_handler.py
@@ -1,0 +1,36 @@
+"""
+Class for handling Linux signal interrupts
+"""
+import signal
+
+class GracefulInterruptHandler(object):
+    def __init__(self, signals=(signal.SIGINT, signal.SIGTERM)):
+        self.signals = signals
+        self.original_handlers = {}
+
+    def __enter__(self):
+        self.interrupted = False
+        self.released = False
+
+        for sig in self.signals:
+            self.original_handlers[sig] = signal.getsignal(sig)
+            signal.signal(sig, self.handler)
+
+        return self
+
+    def handler(self, signum, frame):
+        self.release()
+        self.interrupted = True
+
+    def __exit__(self, type, value, tb):
+        self.release()
+
+    def release(self):
+        if self.released:
+            return False
+
+        for sig in self.signals:
+            signal.signal(sig, self.original_handlers[sig])
+
+        self.released = True
+        return True
